[WIP] tests for checkProgressOnBranch
