+++ lib/utilities/crowdin-api.js
@@ -7,26 +7,26 @@ const config = require('./config');
 
 class CrowdinApi {
   static buildBranch(crowdinBranchName) {
-    return this._crowdinFetch(`export`, { branch: crowdinBranchName });
+    return _crowdinFetch(`export`, { branch: crowdinBranchName });
   }
 
   static getAllTranslations(crowdinBranchName) {
-    return this._crowdinFetch(`download/all.zip`, { branch: crowdinBranchName }, false);
+    return _crowdinFetch(`download/all.zip`, { branch: crowdinBranchName }, false);
   }
 
   static addBranch(crowdinBranchName) {
-    return this._crowdinFetch('add-directory', {
+    return _crowdinFetch('add-directory', {
       name: crowdinBranchName,
       is_branch: '1'
     })
   }
 
   static deleteBranch(crowdinBranchName) {
-    return this._crowdinFetch('delete-directory', { name: crowdinBranchName });
+    return _crowdinFetch('delete-directory', { name: crowdinBranchName });
   }
 
   static addDirectory(name, crowdinBranchName) {
-    return this._crowdinFetch('add-directory', {
+    return _crowdinFetch('add-directory', {
       name,
       branch: crowdinBranchName,
       recursive: '1'
@@ -34,7 +34,7 @@ class CrowdinApi {
   }
 
   static addFile(filePath, crowdinBranchName) {
-    return this._crowdinFetch('add-file', {
+    return _crowdinFetch('add-file', {
       ['files[' + filePath + ']']: fs.createReadStream(filePath),
       ['export_patterns[' + filePath + ']']: config.translationPattern,
       branch: crowdinBranchName
@@ -42,7 +42,7 @@ class CrowdinApi {
   }
 
   static updateFile(filePath, crowdinBranchName) {
-    return this._crowdinFetch('update-file', {
+    return _crowdinFetch('update-file', {
       ['files[' + filePath + ']']: fs.createReadStream(filePath),
       ['export_patterns[' + filePath + ']']: config.translationPattern,
       branch: crowdinBranchName
@@ -50,11 +50,11 @@ class CrowdinApi {
   }
 
   static getLanguageStatus() {
-    return this._crowdinFetch('language-status', { language: config.languageToCheck });
+    return _crowdinFetch('language-status', { language: config.languageToCheck });
   }
 
   static preTranslate(sourceFilesWithBranchName) {
-    return this._crowdinFetch('pre-translate', {
+    return _crowdinFetch('pre-translate', {
       'languages': config.languagesToAutoTranslate,
       'files': sourceFilesWithBranchName,
       'method': 'tm',
@@ -64,35 +64,36 @@ class CrowdinApi {
   }
 
   static getInfo() {
-    return this._crowdinFetch('info');
-  }
-
-
-  static async _crowdinFetch(apiMethod, rawParams = {}, isJsonResponse = true) {
-    const formData = new FormData();
-
-    formData.append('key', config.projectKey);
-    formData.append('json', '');
-
-    for (const key in rawParams) {
-      if (Array.isArray(rawParams[key])) {
-        rawParams[key].forEach((value) => {
-          formData.append(`${ key }[]`, value);
-        });
-      } else {
-        formData.append(key, rawParams[key]);
-      }
-    }
-
-    const response = await fetch(
-      `https://api.crowdin.com/api/project/${ config.projectIdentifier }/${ apiMethod }`,
-      { method: 'POST', body: formData }
-    );
-
-    return isJsonResponse
-      ? response.json()
-      : response;
+    return _crowdinFetch('info');
   }
 }
+
+
+async function _crowdinFetch(apiMethod, rawParams = {}, isJsonResponse = true) {
+  const formData = new FormData();
+
+  formData.append('key', config.projectKey);
+  formData.append('json', '');
+
+  for (const key in rawParams) {
+    if (Array.isArray(rawParams[key])) {
+      rawParams[key].forEach((value) => {
+        formData.append(`${ key }[]`, value);
+      });
+    } else {
+      formData.append(key, rawParams[key]);
+    }
+  }
+
+  const response = await fetch(
+    `https://api.crowdin.com/api/project/${ config.projectIdentifier }/${ apiMethod }`,
+    { method: 'POST', body: formData }
+  );
+
+  return isJsonResponse
+    ? response.json()
+    : response;
+}
+
 
 module.exports = CrowdinApi;
